separate napari-visible functions from under-the-hood functions
