+++ src/napari_stress/_spherical_harmonics/spherical_harmonics_napari.py
@@ -1,21 +1,67 @@
 # -*- coding: utf-8 -*-
-
-from napari.types import PointsData, LayerDataTuple
+from napari.types import LayerDataTuple, PointsData
+import numpy as np
+from enum import Enum
 
 from .._utils.frame_by_frame import frame_by_frame
-
+from .spherical_harmonics import shtools_spherical_harmonics_expansion,\
+    stress_spherical_harmonics_expansion,\
+    integrate_on_manifold
+from . import lebedev_info_SPB as lebedev_info
 from . import sph_func_SPB as sph_f
 from . import euclidian_k_form_SPB as euc_kf
-from . import lebedev_info_SPB as lebedev_info
-from . import manifold_SPB as mnfd
-from .expansion import spherical_harmonics_methods
 
-import numpy as np
 import napari
+import warnings
 from napari_tools_menu import register_function
 
-import numpy as np
-import warnings
+
+class spherical_harmonics_methods(Enum):
+    """Available methods for spherical harmonics expansion."""
+
+    shtools = {'function': shtools_spherical_harmonics_expansion}
+    stress = {'function': stress_spherical_harmonics_expansion}
+
+@register_function(menu="Points > Fit spherical harmonics (n-STRESS")
+@frame_by_frame
+def fit_spherical_harmonics(points: PointsData,
+                            max_degree: int = 5,
+                            implementation: spherical_harmonics_methods = spherical_harmonics_methods.stress
+                            ) -> LayerDataTuple:
+    """
+    Approximate a surface by spherical harmonics expansion.
+
+    Parameters
+    ----------
+    points : PointsData
+    max_degree : int
+        Order up to which spherical harmonics should be included for the approximation.
+
+    Returns
+    -------
+    PointsData
+        Pointcloud on surface of a spherical harmonics expansion at the same
+        latitude/longitude as the input points.
+
+    See Also
+    --------
+    [1] https://en.wikipedia.org/wiki/Spherical_harmonics#/media/File:Spherical_Harmonics.png
+
+    """
+    # Parse inputs
+    if isinstance(implementation, str):
+        fit_function = spherical_harmonics_methods.__members__[implementation].value['function']
+    else:
+        fit_function = implementation.value['function']
+    fitted_points, coefficients = fit_function(points, max_degree=max_degree)
+
+    properties, features = {}, {}
+    features['error'] = np.linalg.norm(fitted_points - points, axis=1)
+    properties['features'] = features
+    properties['face_color'] = 'error'
+    properties['size'] = 0.5
+
+    return (fitted_points, properties, 'points')
 
 @register_function(menu="Measurement > Surface curvature from points (n-STRESS)",
                    number_of_quadrature_points={'min': 6, 'max': 5180})
@@ -82,7 +128,7 @@ def measure_curvature(points: PointsData,
     lebedev_points = np.stack(lebedev_points).squeeze().transpose()
 
     properties, features = {}, {}
-    features['curvature'] = _integrate_on_manifold(lebedev_points, LBDV_Fit, max_degree).squeeze()
+    features['curvature'] = integrate_on_manifold(lebedev_points, LBDV_Fit, max_degree).squeeze()
 
     properties['features'] = features
     properties['face_color'] = 'curvature'
@@ -97,39 +143,3 @@ def measure_curvature(points: PointsData,
             layer.features = features
             layer.data = lebedev_points
     return lebedev_points
-
-
-def _integrate_on_manifold(lebedev_points: PointsData, LBDV_Fit, max_degree: int):
-
-    # create manifold to calculate H, average H:
-    Manny_Dict = {}
-    Manny_Name_Dict = {} # sph point cloud at lbdv
-    Manny_Name_Dict['coordinates'] = lebedev_points
-
-    Manny_Dict['Pickle_Manny_Data'] = False
-    Manny_Dict['Maniold_lbdv'] = LBDV_Fit
-
-    Manny_Dict['Manifold_SPH_deg'] = max_degree
-    Manny_Dict['use_manifold_name'] = False # we are NOT using named shapes in these tests
-    Manny_Dict['Maniold_Name_Dict'] = Manny_Name_Dict # sph point cloud at lbdv
-
-    Manny = mnfd.manifold(Manny_Dict)
-
-    # Test orientation:
-    centered_lbdv_pts = lebedev_points - lebedev_points.mean(axis=0)[None, :]
-
-    normal_X_lbdv_pts = euc_kf.Combine_Chart_Quad_Vals(Manny.Normal_Vec_X_A_Pts, Manny.Normal_Vec_X_B_Pts, LBDV_Fit)
-    normal_Y_lbdv_pts = euc_kf.Combine_Chart_Quad_Vals(Manny.Normal_Vec_Y_A_Pts, Manny.Normal_Vec_Y_B_Pts, LBDV_Fit)
-    normal_Z_lbdv_pts = euc_kf.Combine_Chart_Quad_Vals(Manny.Normal_Vec_Z_A_Pts, Manny.Normal_Vec_Z_B_Pts, LBDV_Fit)
-
-    normals_lbdv_points = np.stack([normal_X_lbdv_pts, normal_Y_lbdv_pts, normal_Z_lbdv_pts]).squeeze().transpose()
-
-    # Makre sure orientation is inward, so H is positive (for Ellipsoid, and small deviations):
-    Orientations = [np.dot(x, y) for x, y in zip(centered_lbdv_pts,  normals_lbdv_points)]
-    num_pos_orr = np.sum(np.asarray(Orientations).flatten() > 0)
-
-    Orientation = 1. # unchanged (we want INWARD)
-    if(num_pos_orr > .5 * len(centered_lbdv_pts)):
-        Orientation = -1.
-
-    return Orientation*euc_kf.Combine_Chart_Quad_Vals(Manny.H_A_pts, Manny.H_B_pts, LBDV_Fit)
